# ENH use more blas functions in cd solvers