#### Reference Issues/PRs
Towards #13210.


#### What does this implement/fix? Explain your changes.
This PR tries to use more BLAS functions in the coordinate descent solvers in Cython.

#### Any other comments?
I did some benchmarking for different options for the computation of `XtA` in `enet_coordinate_descent_multi_task` without clear conclusion. Therefore, I let it be as is with more comments added.

<details>

```python
from collections import OrderedDict
from itertools import product
import time

from neurtu import delayed, timeit, Benchmark
import numpy as np
import pandas as pd

from sklearn.datasets import make_regression
from sklearn.linear_model import MultiTaskLasso


alpha = 0.01
list_n_features = [300, 1000, 4000]
list_n_samples = [100, 500]
list_n_tasks = [2, 10, 20, 50]

X, Y, coef_ = make_regression(
    n_samples=max(list_n_samples),
    n_features=max(list_n_features),
    n_targets=max(list_n_tasks),
    n_informative=max(list_n_features) // 10,
    noise=0.1,
    coef=True
)

X /= np.sqrt(np.sum(X ** 2, axis=0))  # Normalize data


def benchmark_cases(X, Y):
    """Benchmark MultiTaskLasso"""
    for it, (ns, nf, nt) in enumerate(product(list_n_samples, list_n_features, list_n_tasks)):
        tags = OrderedDict(n_samples=ns, n_features=nf, n_targets=nt)
        clf = MultiTaskLasso(alpha=alpha, fit_intercept=False, max_iter=10_000)
        yield delayed(clf.fit, tags=tags)(X[:ns, :nf], Y[:ns, :nt])

bench = Benchmark(repeat=5)
print("Run benchmark for multi target lasso.")
start = time.time()
result = bench(benchmark_cases(X, Y))
end =  time.time()
print(f"Finished after {end - start} seconds.")
print(result)
```

</details>