@lorentzenchr 2022-03-28
Benchmark with the scipt in the top post based on https://github.com/scikit-learn/scikit-learn/pull/22972/commits/7623065e09430a906911a83056a6288160ba00dd. Speedup is main / PR.
```
PR with GEMM
                                 wall_time                      
                                      mean         max       std  speedup
n_samples n_features n_targets                                  
100       300        2            0.145484    0.147704  0.001475  1.08
                     10           0.677363    0.684821  0.010156  0.955
                     20           1.496502    1.514316  0.020742  0.927
                     50           5.971845    6.008925  0.038615  0.594
          1000       2            0.497537    0.509534  0.008607  1.27
                     10           2.488258    2.514491  0.017274  1.58
                     20           6.157018    6.181837  0.023664  1.31
                     50          20.179273   20.565162  0.235771  1.06
          4000       2            2.141750    2.175920  0.032265  1.36
                     10          14.191492   14.516755  0.186170  1.06
                     20          26.280743   26.790688  0.296696  1.13
                     50          71.058634   71.500954  0.315243  1.09
500       300        2            0.008196    0.008323  0.000087  1.26
                     10           0.041458    0.042422  0.000625  1.24
                     20           0.086655    0.095563  0.004997  1.24
                     50           0.238273    0.262718  0.015520  1.34
          1000       2            1.162102    1.173390  0.010608  1.11
                     10           5.397492    5.425889  0.021651  1.16
                     20          12.493163   12.701867  0.122206  1.18
                     50          35.322965   36.820687  1.244409  1.57
          4000       2            4.084729    4.101981  0.012868  1.03
                     10          21.156687   21.183405  0.017483  0.86
                     20          44.667367   44.732523  0.059359  0.97
                     50         143.688414  147.081807  4.331627  1.07

```
Timings of main branch
<details>

```
MAIN
                                 wall_time                      
                                      mean         max       std
n_samples n_features n_targets                                  
100       300        2            0.156672    0.186633  0.017599
                     10           0.647087    0.729176  0.049114
                     20           1.387310    1.493217  0.072670
                     50           3.546491    3.717894  0.106801
          1000       2            0.634045    0.688953  0.040473
                     10           3.937150    4.127279  0.126444
                     20           8.093542    8.752413  0.439550
                     50          21.394161   22.490225  0.808843
          4000       2            2.920357    3.397627  0.286309
                     10          14.978383   16.277204  0.920301
                     20          29.779723   31.012935  1.112678
                     50          77.519848   78.394667  1.187355
500       300        2            0.010294    0.012238  0.001227
                     10           0.051599    0.053901  0.001610
                     20           0.107360    0.116611  0.008979
                     50           0.318203    0.358433  0.035147
          1000       2            1.293451    1.574638  0.202206
                     10           6.266583    7.150905  0.760419
                     20          14.796782   16.219736  0.969471
                     50          55.599680   59.203180  3.049441
          4000       2            4.200956    4.325626  0.127612
                     10          18.290157   18.776878  0.340980
                     20          43.271756   46.699309  2.054371
                     50         153.149346  158.434174  3.035532
```

</details>