@lorentzenchr 2022-04-10
> What does sklearn.show_versions() look like for you?

```python
import sklearn
sklearn.show_versions()
```

<details>

```
System:
    python: 3.9.9 (main, Nov 21 2021, 03:23:44)  [Clang 13.0.0 (clang-1300.0.29.3)]
executable: /Users/lorentzen/github/python3_sklearn/bin/python
   machine: macOS-12.3-x86_64-i386-64bit

Python dependencies:
      sklearn: 1.1.dev0
          pip: 22.0.4
   setuptools: 58.1.0
        numpy: 1.21.4
        scipy: 1.7.3
       Cython: 0.29.24
       pandas: 1.3.4
   matplotlib: 3.5.0
       joblib: 1.1.0
threadpoolctl: 3.1.0

Built with OpenMP: True

threadpoolctl info:
       user_api: openmp
   internal_api: openmp
         prefix: libomp
       filepath: /usr/local/Cellar/libomp/13.0.0/lib/libomp.dylib
        version: None
    num_threads: 8

       user_api: blas
   internal_api: openblas
         prefix: libopenblas
       filepath: /Users/lorentzen/github/python3_sklearn/lib/python3.9/site-packages/numpy/.dylibs/libopenblas.0.dylib
        version: 0.3.17
threading_layer: pthreads
   architecture: Haswell
    num_threads: 4

       user_api: blas
   internal_api: openblas
         prefix: libopenblas
       filepath: /Users/lorentzen/github/python3_sklearn/lib/python3.9/site-packages/scipy/.dylibs/libopenblas.0.dylib
        version: 0.3.17
threading_layer: pthreads
   architecture: Haswell
    num_threads: 4
```

</details>