@lorentzenchr 2022-04-13
> Can you provide a benchmark showing that there is no regression in performance?

See https://github.com/scikit-learn/scikit-learn/pull/22972#issuecomment-1081096256. TLDR: It's a trade-off.